errors fixed
