+++ .dart_tool/package_config_subset
@@ -54,6 +54,10 @@ ffi
 2.17
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/ffi-2.0.1/
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/ffi-2.0.1/lib/
+file
+2.12
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/file-6.1.4/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/file-6.1.4/lib/
 file_picker
 2.17
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/file_picker-5.2.11/
@@ -66,6 +70,10 @@ flutter_plugin_android_lifecycle
 2.12
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.9/
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.9/lib/
+google_fonts
+2.14
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/google_fonts-4.0.4/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/google_fonts-4.0.4/lib/
 http
 2.14
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/http-0.13.5/
@@ -98,10 +106,38 @@ path
 2.12
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path-1.8.2/
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path-1.8.2/lib/
+path_provider
+2.17
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider-2.0.15/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider-2.0.15/lib/
+path_provider_android
+2.17
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_android-2.0.27/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_android-2.0.27/lib/
+path_provider_foundation
+2.18
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_foundation-2.2.3/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_foundation-2.2.3/lib/
+path_provider_linux
+2.17
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_linux-2.1.10/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_linux-2.1.10/lib/
+path_provider_platform_interface
+2.12
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_platform_interface-2.0.6/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_platform_interface-2.0.6/lib/
+path_provider_windows
+2.17
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_windows-2.1.6/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_windows-2.1.6/lib/
 petitparser
 2.18
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/petitparser-5.1.0/
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/petitparser-5.1.0/lib/
+platform
+2.12
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/platform-3.1.0/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/platform-3.1.0/lib/
 plugin_platform_interface
 2.17
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/plugin_platform_interface-2.1.4/
@@ -110,6 +146,10 @@ pointycastle
 2.14
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/pointycastle-3.7.3/
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/pointycastle-3.7.3/lib/
+process
+2.14
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/process-4.2.4/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/process-4.2.4/lib/
 source_span
 2.14
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/source_span-1.9.1/
@@ -134,6 +174,10 @@ win32
 2.17
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/win32-4.1.3/
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/win32-4.1.3/lib/
+xdg_directories
+2.17
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/xdg_directories-1.0.0/
+file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/xdg_directories-1.0.0/lib/
 xml
 2.18
 file:///C:/Users/HP/AppData/Local/Pub/Cache/hosted/pub.dev/xml-6.2.2/
