+++ lib/submit_assignment/assignment.dart
@@ -1,5 +1,5 @@
 import 'dart:io';
-import 'package:google_fonts/google_fonts.dart';
+//import 'package:google_fonts/google_fonts.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_application_5/Src/widgets/sideNav.dart';
